fix(smart-wallet): move types to a separate file
