+++ packages/client/wallets/smart-wallet/src/apiService.ts
@@ -1,227 +1,13 @@
 import type { BlockchainIncludingTestnet as Blockchain, Crossmint } from "@crossmint/common-sdk-base";
 import { CrossmintApiClient } from "@crossmint/common-sdk-base";
 import { APIErrorService, SDKLogger } from "@crossmint/client-sdk-base";
-import type { Address, Hex, TypedData, TypedDataDomain } from "viem";
-
-import type { SmartWalletChain } from "@/evm/chains";
+import type { Address } from "viem";
 
 import { SDK_NAME, SDK_VERSION, SCW_SERVICE, API_VERSION } from "./utils/constants";
 import type { UserParams } from "./smartWalletService";
+import type { CreateWalletParams, CreateWalletResponse, CreateTransactionParams, TransactionResponse, ApproveTransactionParams, CreateSignatureParams, SignatureResponse, ApproveSignatureParams, NftResponse } from "./types/api";
 
 export const scwLogger = new SDKLogger(SCW_SERVICE);
-
-export type Signer = `evm-keypair:${Address}` | `evm-passkey:${string}`;
-
-interface CreateWalletParams {
-    type: "evm-smart-wallet";
-    config: {
-        adminSigner:
-            | {
-                  type: "evm-keypair";
-                  address: Address;
-              }
-            | {
-                  type: "evm-passkey";
-                  id: string;
-                  name: string;
-                  publicKey: {
-                      x: Hex;
-                      y: Hex;
-                  };
-              };
-        creationSeed: "0";
-    };
-}
-
-export interface CreateWalletResponse {
-    type: "evm-smart-wallet";
-    linkedUser: string;
-    address: Address;
-    config: {
-        adminSigner:
-            | {
-                  type: "evm-keypair";
-                  locator: `evm-keypair:${Address}`;
-                  address: Address;
-              }
-            | {
-                  type: "evm-passkey";
-                  locator: `evm-passkey:${string}`;
-                  id: string;
-                  name: string;
-                  publicKey: {
-                      x: Hex;
-                      y: Hex;
-                  };
-                  validatorContractVersion: string;
-              };
-    };
-}
-
-interface TransactionCall {
-    to: Address;
-    value: string;
-    data: Hex;
-}
-
-interface CreateTransactionParams {
-    params: {
-        signer: Signer;
-        chain: SmartWalletChain;
-        calls: TransactionCall[];
-    };
-}
-
-export interface TransactionResponse {
-    id: string;
-    walletType: "evm-smart-wallet";
-    status: "awaiting-approval" | "pending" | "failed" | "success";
-    approvals: {
-        pending: {
-            message: Hex;
-            signer: Signer;
-        }[];
-        submitted: {
-            message: Hex;
-            signature: Hex;
-            signer: Signer;
-        }[];
-    };
-    onChain: {
-        userOperationHash: Hex;
-        txId?: Hex;
-    };
-    error?: {
-        message: string;
-        reason: 'execution_reverted';
-        revert: {
-            type: 'contract_call' | 'wallet_authorization' | 'wallet_deployment';
-            reason: string;
-            reasonData?: string;
-            explorerLink?: string;   
-            simulationLink?: string;
-        }
-    };
-}
-
-type Approval =
-    | {
-          signer: `evm-keypair:${Address}`;
-          signature: Hex;
-      }
-    | {
-          signer: `evm-passkey:${string}`;
-          signature: {
-              r: Hex;
-              s: Hex;
-          };
-          metadata: {
-              authenticatorData: Hex;
-              challengeIndex: number;
-              clientDataJSON: string;
-              typeIndex: number;
-              userVerificationRequired: boolean;
-          };
-      };
-
-interface ApproveTransactionParams {
-    approvals: Approval[];
-}
-
-type CreateSignatureParams =
-    | {
-          type: "evm-message";
-          params: {
-              message: string;
-              signer: Signer;
-              chain: SmartWalletChain;
-          };
-      }
-    | {
-          type: "evm-typed-data";
-          params: {
-              chain: SmartWalletChain;
-              signer: Signer;
-              isSmartWalletSignature: boolean;
-              typedData: {
-                  domain: TypedDataDomain;
-                  message: Record<string, unknown>;
-                  primaryType: string;
-                  types: TypedData;
-              };
-          };
-      };
-
-interface SignatureResponseBase {
-    id: string;
-    walletType: "evm-smart-wallet";
-    status: "awaiting-approval" | "pending" | "failed" | "success";
-    approvals: {
-        pending: {
-            signer: Signer;
-            message: Hex;
-        }[];
-        submitted: {
-            signer: Signer;
-            signature: Hex;
-            message: Hex;
-        }[];
-    };
-    createdAt: string;
-}
-
-interface ApproveSignatureParams {
-    approvals: Approval[];
-}
-
-interface SignatureResponseMessage extends SignatureResponseBase {
-    type: "evm-message";
-    params: {
-        message: string;
-        signer: Signer;
-        chain: SmartWalletChain;
-    };
-}
-
-interface SignatureResponseTypedData extends SignatureResponseBase {
-    type: "evm-typed-data";
-    params: {
-        chain: SmartWalletChain;
-        signer: Signer;
-        isSmartWalletSignature: boolean;
-        typedData: {
-            domain: TypedDataDomain;
-            message: Record<string, unknown>;
-            primaryType: string;
-            types: Record<string, unknown>;
-        };
-    };
-}
-
-export type SignatureResponse = SignatureResponseMessage | SignatureResponseTypedData;
-
-type NftResponse = {
-    chain: Blockchain;
-    contractAddress: Address;
-    tokenId: string;
-    metadata: {
-        attributes: {
-            display_type: string | null;
-            trait_type: string;
-            value: string;
-        }[];
-        collection: unknown;
-        description: string;
-        image: string;
-        animation_url: string | null;
-        name: string;
-    };
-    subscription?: {
-        expiresAt: string;
-    };
-    locator: string;
-    tokenStandard: string;
-}[];
 
 type WalletsAPIErrorCodes = never;
