+++ packages/client/wallets/smart-wallet/src/smartWalletService.ts
@@ -22,11 +22,13 @@ import entryPointAbi from "@/abi/entryPoint";
 
 import type {
     CrossmintApiService,
+} from "./apiService";
+import type {
     CreateWalletResponse,
     TransactionResponse,
     Signer,
     SignatureResponse,
-} from "./apiService";
+} from "./types/api";
 import type { SmartWalletChain } from "./evm/chains";
 import type { SmartWalletClient } from "./evm/smartWalletClient";
 import { EVMSmartWallet } from "./evm/wallet";
